serving/requirements.txt
google-cloud-storage
fastapi
uvicorn
joblib
numpy
pandas
lightgbm
scikit-learn
